updates and changed component_attachment to attribute

+++ datasets/NHS-guardianship-mental-health-act/table4.py
@@ -99,7 +99,7 @@ new_table = new_table.replace({'Local authority name' : {'' : 'entire-region'}})
 new_table = new_table.replace({'Local authority code' : {'-' : 'not-applicable'}})
 new_table = new_table.fillna('not-applicable')
 
-new_table['Median months'] = new_table['Median months'].astype(float)
+#new_table['Median months'] = new_table['Median months'].astype(float)
 # -
